requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pymupdf>=1.23.0
python-dotenv>=1.0.0
nltk>=3.8.0
